Create model views
When trying to get a trip's detail, users can only use the 'trip_id' instead of using the 'id', since they can easily just enter a number to view details on trips. This also allows the front end, to use the trip id that a user would email to their friends once they create a trip.